SMA-12: Login and Test page now display when respective URL entered, sample code for SportEvent moved to Home page
