+++ frontend/sportsmatch-app/src/pages/Login.tsx
@@ -1,0 +1,3 @@
+export default function Login() {
+  return <h1>This is a login page</h1>;
+}
