+++ frontend/sportsmatch-app/src/pages/Test.tsx
@@ -1,0 +1,7 @@
+export default function Test() {
+  return (
+    <div>
+      <h1>Test</h1>
+    </div>
+  );
+}
